requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
python-dotenv==0.19.0
pyjwt==2.1.0
werkzeug==2.0.1
gunicorn==20.1.0
flask-restx==1.1.0
flask-jwt-extended==4.7.1
bcrypt==4.0.1 